requirements.txt
aiogram==2.25.1

requests==2.31.0

python-dotenv==1.0.0

Flask==2.3.2